# Dev/improve performance in do sidetable control transfer